Eliminate g++ -Wreorder warnings
Rearrange declaration order to be identical to member initialization order.
Different order causes -Wreorder warning if compiling with g++ and might in general cause other side effects.
In this case no errors actually occur, but changing order seems to be the better solution then turning off warning for the whole project.
More information can be found on https://stackoverflow.com/questions/1828037/whats-the-point-of-g-wreorder.