Debug ServerTP::ThreadStarter() to allow multiple starts and stops of the thread pool
Previously the problem was:
If a StopThread() is called for a ServerTP, the ThreadStarter() function joins all the worker threads. However the std::vector<std::thread> workers_ is not cleared, it now consists of empty threads. If you restart the server, the vector will contain the old, invalid threads and the new ones. On the next StopThread() the worker.join() on the old threads throws an uncaught exception.