Add "unused" attribute to #define log_define(category)
According to commentary, the define should be used to "define the logger that will be used in the current namespace or class."
However using the log_define in an namespace (of a header file) will lead to "-Wunused-function" warnings with the gcc compiler,
if not all including cpp-files actually print a log message. This is especially troublesome if compiling with "-Werror".

Using the __attribute__ ((unused)), this warning is disabled.
Using the ATTR_UNUSED define-wrapper prevents other compilers from being affected.